Solve 136 and add header for 137.

+++ LeetCode/135_Candy.h
@@ -16,7 +16,7 @@ public:
             }
         }
         
-        for (int i = ratings.size() -2; i >= 0; i--)
+        for (int i = static_cast<int>(ratings.size() -2); i >= 0; i--)
         {
             if (ratings[i] > ratings[i + 1])
             {
